+++ checkpoint/orbax/checkpoint/checkpoint_utils.py
@@ -13,6 +13,7 @@
 # limitations under the License.
 
 """High-level checkpoint utils provided for user convenience."""
+
 import contextlib
 import os
 import time
@@ -23,6 +24,7 @@ from etils import epath
 import jax
 from jax.experimental import multihost_utils
 import numpy as np
+from orbax.checkpoint import storage
 from orbax.checkpoint import type_handlers
 from orbax.checkpoint import utils
 from orbax.checkpoint import value_metadata
@@ -32,22 +34,27 @@ PyTree = Any
 STANDARD_ARRAY_TYPES = (int, float, np.ndarray, jax.Array)
 
 
-def _lock_checkpoint(
-    checkpoint_dir: epath.Path,
-    step: int,
-    step_prefix: Optional[str],
-    step_format_fixed_length: Optional[int],
-) -> bool:
-  """Locks a checkpoint by writing a LOCKED directory."""
-  directory = utils.get_save_directory(
-      step,
-      checkpoint_dir,
+def _init_step_lookup(
+    step_lookup: Optional[storage.StepLookup] = None,
+    step_prefix: Optional[str] = None,
+    step_format_fixed_length: Optional[int] = None,
+):
+  return step_lookup or storage.DefaultStepLookup(
       step_prefix=step_prefix,
       step_format_fixed_length=step_format_fixed_length,
   )
-  if not directory.exists():
-    raise ValueError(f'Parent directory {directory} does not exist.')
-  lockdir = utils.lockdir(directory)
+
+
+def _lock_checkpoint(
+    checkpoint_dir: epath.Path,
+    step: int,
+    step_lookup: storage.StepLookup,
+) -> bool:
+  """Locks a checkpoint by writing a LOCKED directory."""
+  step_dir = storage.root(checkpoint_dir).lookup(step, step_lookup).path
+  if not step_dir.exists():
+    raise ValueError(f'Step directory {step_dir} does not exist.')
+  lockdir = utils.lockdir(step_dir)
   try:
     lockdir.mkdir(parents=False, exist_ok=True)
     return True
@@ -64,38 +71,40 @@ def _lock_checkpoint(
 def _unlock_checkpoint(
     checkpoint_dir: epath.Path,
     step: int,
-    step_prefix: Optional[str],
-    step_format_fixed_length: Optional[int],
+    step_lookup: storage.StepLookup,
 ):
   """Removes a LOCKED directory to indicate unlocking."""
   if jax.process_index() == 0:
-    directory = utils.get_save_directory(
-        step,
-        checkpoint_dir,
-        step_prefix=step_prefix,
-        step_format_fixed_length=step_format_fixed_length,
-    )
-    utils.lockdir(directory).unlink(missing_ok=True)
+    step_dir = storage.root(checkpoint_dir).lookup(step, step_lookup).path
+    utils.lockdir(step_dir).unlink(missing_ok=True)
 
 
 def unlock_existing_checkpoints(
     checkpoint_dir: epath.Path,
-    step_prefix: Optional[str],
-    step_format_fixed_length: Optional[int],
+    step_prefix: Optional[str] = None,
+    step_format_fixed_length: Optional[int] = None,
+    step_lookup: Optional[storage.StepLookup] = None,
 ):
-  """Removes LOCKED file for all existing steps, if present."""
+  """Removes LOCKED file for all existing steps, if present.
+
+  Args:
+    checkpoint_dir: The directory containing StepDirs.
+    step_prefix: A prefix applied to step numbers (e.g. <prefix>_42).
+    step_format_fixed_length: Expects to find checkpoint step directories with
+      exactly this number of digits (leading zeros if necessary).
+    step_lookup: StepLookup used to lookup step under given root directory. If
+      provided, `step_prefix` and `step_format_fixed_length` are ignored.
+  """
   steps = utils.checkpoint_steps(checkpoint_dir)
+  step_lookup = _init_step_lookup(
+      step_lookup=step_lookup,
+      step_prefix=step_prefix,
+      step_format_fixed_length=step_format_fixed_length,
+  )
   for step in steps:
-    directory = utils.get_save_directory(
-        step,
-        checkpoint_dir,
-        step_prefix=step_prefix,
-        step_format_fixed_length=step_format_fixed_length,
-    )
-    if utils.is_locked(directory):
-      _unlock_checkpoint(
-          checkpoint_dir, step, step_prefix, step_format_fixed_length
-      )
+    step_dir = storage.root(checkpoint_dir).lookup(step, step_lookup).path
+    if utils.is_locked(step_dir):
+      _unlock_checkpoint(checkpoint_dir, step, step_lookup)
 
 
 def checkpoint_dir_is_world_readable(
@@ -120,11 +129,10 @@ def _reached_desired_step(step: int, until_step: Optional[int]) -> bool:
 
 def _wait_for_new_checkpoint(
     checkpoint_dir: epath.Path,
+    step_lookup: storage.StepLookup,
     until_step: Optional[int] = None,
     seconds_to_sleep: int = 1,
     timeout: Optional[int] = None,
-    step_prefix: Optional[str] = None,
-    step_format_fixed_length: Optional[int] = None,
 ) -> int:
   """See documentation for wait_for_new_checkpoint."""
   start = time.time()
@@ -153,12 +161,7 @@ def _wait_for_new_checkpoint(
       steps = utils.checkpoint_steps(checkpoint_dir)
       checkpoint_step = max(steps) if steps else None
       if _reached_desired_step(checkpoint_step, until_step):
-        if not _lock_checkpoint(
-            checkpoint_dir,
-            checkpoint_step,
-            step_prefix,
-            step_format_fixed_length,
-        ):
+        if not _lock_checkpoint(checkpoint_dir, checkpoint_step, step_lookup):
           continue
         result = checkpoint_step
         break
@@ -186,6 +189,7 @@ def wait_for_new_checkpoint(
     timeout: Optional[int] = None,
     step_prefix: Optional[str] = None,
     step_format_fixed_length: Optional[int] = None,
+    step_lookup: Optional[storage.StepLookup] = None,
 ):
   """Waits until a new checkpoint file is found.
 
@@ -204,26 +208,30 @@ def wait_for_new_checkpoint(
     step_prefix: A prefix applied to step numbers (e.g. <prefix>_42).
     step_format_fixed_length: Expects to find checkpoint step directories with
       exactly this number of digits (leading zeros if necessary).
+    step_lookup: StepLookup used to lookup step under given root directory. If
+      provided, `step_prefix` and `step_format_fixed_length` are ignored.
 
   Yields:
     a new checkpoint step, or -1 if the timeout was reached.
   """
+  step_lookup = _init_step_lookup(
+      step_lookup=step_lookup,
+      step_prefix=step_prefix,
+      step_format_fixed_length=step_format_fixed_length,
+  )
   step = _wait_for_new_checkpoint(
       checkpoint_dir,
+      step_lookup,
       until_step=until_step,
       seconds_to_sleep=seconds_to_sleep,
       timeout=timeout,
-      step_prefix=step_prefix,
-      step_format_fixed_length=step_format_fixed_length,
   )
   try:
     yield step
   finally:
     # Release lock on the checkpoint step.
     if step != -1:
-      _unlock_checkpoint(
-          checkpoint_dir, step, step_prefix, step_format_fixed_length
-      )
+      _unlock_checkpoint(checkpoint_dir, step, step_lookup)
 
 
 def checkpoints_iterator(
@@ -233,6 +241,7 @@ def checkpoints_iterator(
     timeout_fn: Optional[Callable[[], bool]] = None,
     step_prefix: Optional[str] = None,
     step_format_fixed_length: Optional[int] = None,
+    step_lookup: Optional[storage.StepLookup] = None,
 ) -> Iterator[int]:
   """Continuously yield new checkpoint files as they appear.
 
@@ -278,15 +287,20 @@ def checkpoints_iterator(
     step_prefix: A prefix applied to step numbers (e.g. <prefix>_42).
     step_format_fixed_length: Expects to find checkpoint step directories with
       exactly this number of digits (leading zeros if necessary).
+    step_lookup: StepLookup used to lookup step under given root directory. If
+      provided, `step_prefix` and `step_format_fixed_length` are ignored.
 
   Yields:
     Integer step numbers of the latest checkpoints as they arrive.
   """
   checkpoint_dir = epath.Path(checkpoint_dir)
+  step_lookup = _init_step_lookup(
+      step_lookup=step_lookup,
+      step_prefix=step_prefix,
+      step_format_fixed_length=step_format_fixed_length,
+  )
   try:
-    unlock_existing_checkpoints(
-        checkpoint_dir, step_prefix, step_format_fixed_length
-    )
+    unlock_existing_checkpoints(checkpoint_dir, step_lookup=step_lookup)
   except FileNotFoundError as e:
     logging.warning(
         'Encountered error while unlocking existing checkpoints. Some'
@@ -300,8 +314,7 @@ def checkpoints_iterator(
         checkpoint_dir,
         until_step=until_step,
         timeout=timeout,
-        step_prefix=step_prefix,
-        step_format_fixed_length=step_format_fixed_length,
+        step_lookup=step_lookup,
     ) as new_checkpoint_step:
       if new_checkpoint_step == -1:
         if not timeout_fn:
@@ -366,8 +379,8 @@ def construct_restore_args(
 
   Args:
     target: The returned PyTree will match the structure of `target`. `target`
-      may contain `value_metadata.Metadata`, real scalar or array values, or
-      may contain jax.ShapeDtypeStruct.
+      may contain `value_metadata.Metadata`, real scalar or array values, or may
+      contain jax.ShapeDtypeStruct.
     sharding_tree: A PyTree matching `target` which will be used to set the
       restoration sharding. If not provided, sharding will default to the
       shardings specified by `target`.